torque_server/requirements.txt
asgiref==3.3.1
Django==3.1.3
djangorestframework==3.12.2
future==0.18.2
generics==2.0.0
importlib-metadata==2.0.0
iso8601==0.1.13
Markdown==3.3.3
numpy==1.19.4
Pillow==8.0.1
pyserial==3.4
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
wxPython==4.1.0
zipp==3.4.0
